+++ packages/fcl/src/current-user/exec-service/strategies/tab-rpc.js
@@ -2,6 +2,7 @@ import {uid} from "@onflow/util-uid"
 import {tab} from "./utils/tab"
 import {normalizePollingResponse} from "../../normalize/polling-response"
 import {configLens} from "../../../config-utils"
+import {VERSION} from "../../../VERSION"
 
 export function execTabRPC(service, body, opts) {
   return new Promise((resolve, reject) => {
@@ -14,6 +15,7 @@ export function execTabRPC(service, body, opts) {
       async onReady(_, {send}) {
         try {
           send({
+            fclVersion: VERSION,
             type: "FCL:VIEW:READY:RESPONSE",
             body,
             service: {
@@ -27,6 +29,7 @@ export function execTabRPC(service, body, opts) {
             },
           })
           send({
+            fclVersion: VERSION,
             type: "FCL:FRAME:READY:RESPONSE",
             body,
             service: {
@@ -110,7 +113,7 @@ export function execTabRPC(service, body, opts) {
               resolve(resp.data)
               close()
               break
-              
+
             default:
               reject(`Declined: No reason supplied`)
               close()