+++ packages/fcl/src/wallet-utils/index.js
@@ -1,4 +1,4 @@
-export {sendMsgToFCL, close, approve, decline} from "./send-msg-to-fcl.js"
+export {sendMsgToFCL, close, approve, decline, redirect} from "./send-msg-to-fcl.js"
 export {onMessageFromFCL} from "./on-message-from-fcl.js"
 export {encodeMessageFromSignable} from "@onflow/sdk"
 export {CompositeSignature} from "./CompositeSignature.js"