+++ packages/fcl/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
-- 2021-10-12 -- Adds `execExtRPC` strategy for use with browser extension communication and new `WalletUtils.redirect` and `WalletUtils.injectExtService` utils. `WalletUtils.injectExtService` can be used by browser extensions to tell FCL how to authenticate and to be shown as an option in Wallet Discovery.
+- 2021-10-12 -- Adds `execExtRPC` strategy for use with browser extension communication and new `WalletUtils.ready`, `WalletUtils.redirect` and `WalletUtils.injectExtService` utils. `WalletUtils.injectExtService` can be used by browser extensions to tell FCL how to authenticate with it.
 
 ```javascript
 import {WalletUtils} from "@onflow/fcl"